[ FIX ] 익명의 사용자가 채널에 입장할 때, 적절하지 않은 에러 모달이 렌더링되는 에러 해결
Why
- 적절하지 않은 에러 모달이 렌더링되지 않아야 한다.

What
- Auth 컴포넌트를 구현하여, Channel 컴포넌트를 구현하기 전에 유저 인증 로직을 처리하도록 하였다.
