+++ web/src/components/common/Auth/index.js
@@ -1,0 +1,1 @@
+export { default } from './Auth';
