+++ web/src/components/common/index.js
@@ -5,3 +5,4 @@ export { default as LoadingModal } from './LoadingModal';
 export { default as ErrorModal } from './ErrorModal';
 export { default as SmallButton } from './SmallButton';
 export { default as FullScreen } from './FullScreen';
+export { default as Auth } from './Auth';
